+++ static/js/main.js
@@ -79,6 +79,7 @@ jQuery(document).ready(function() {
     const crosshairMode = jQuery(".crosshair-mode");
     const viewportOptionsCrosshair = jQuery(".viewport-options-crosshair");
     const keyboardMode = jQuery(".keyboard-mode");
+    const viewportOptionsKeyboard = jQuery(".viewport-options-keyboard");
     const guidelinesMode = jQuery(".guidelines-mode");
     const viewportOptionsGuidelines = jQuery(".viewport-options-guidelines");
     const caretMode = jQuery(".caret-mode");
@@ -100,6 +101,12 @@ jQuery(document).ready(function() {
     const emojisContainer = jQuery(".emojis-container");
     const emojispContainer = jQuery(".emojisp-container");
     const viewportContainer = jQuery(".viewer-container");
+    const calligraphyContainer = jQuery(".calligraphy-container");
+    const calligraphyMode = jQuery(".calligraphy-mode");
+    const calligraphyModeContainer = jQuery(".calligraphy-mode-container");
+    const calligraphyControls = jQuery(".calligraphy-controls");
+    const calligraphyUndo = jQuery(".calligraphy-undo");
+    const calligraphyClear = jQuery(".calligraphy-clear");
     const formConsole = jQuery(".form-console");
     const inputViewport = jQuery(".input-viewport");
     const signature = jQuery(".signature");
@@ -383,6 +390,13 @@ jQuery(document).ready(function() {
                 viewportContainer.find("> .caret").hide();
                 viewportContainer.removeClass("caret-active");
             }
+
+            // restores the calligraphy mode from the URL
+            // query parameters if it was previously saved
+            const urlCalligraphy = urlParams.get("calligraphy");
+            if (urlCalligraphy === "1" && currentProfile) {
+                calligraphyMode.prop("checked", true).trigger("change");
+            }
             restoring = false;
             updateUrl();
         } catch (err) {
@@ -659,6 +673,7 @@ jQuery(document).ready(function() {
             viewportOptionsCaret.addClass("visible");
             zoomContainer.addClass("visible");
             positionContainer.addClass("visible");
+            calligraphyModeContainer.addClass("visible");
         } else {
             marginContainer.removeClass("visible");
             viewportOptionsRulers.removeClass("visible");
@@ -667,7 +682,14 @@ jQuery(document).ready(function() {
             viewportOptionsCaret.removeClass("visible");
             zoomContainer.removeClass("visible");
             positionContainer.removeClass("visible");
+            calligraphyModeContainer.removeClass("visible");
+            calligraphyControls.removeClass("visible");
         }
+        // resets the calligraphy canvas when switching profiles
+        // since the canvas dimensions change with the profile
+        calligraphyContainer.calligraphy("reset");
+        body.data("calligraphy", null);
+
         const textConfig = currentProfile ? currentProfile.text || {} : {};
         const maxLines = currentProfile ? textConfig.max_lines || 0 : 1;
         viewportContainer.texteditor("option", { maxLines: maxLines });
@@ -678,6 +700,12 @@ jQuery(document).ready(function() {
         updateFontSizeControls(currentProfile);
         applyFontSize();
         inspirationPanel.inspirationpanel("update", currentProfile);
+
+        // re-initializes the calligraphy canvas if the mode
+        // is still active after the profile switch completes
+        if (calligraphyMode.prop("checked") && currentProfile) {
+            initCalligraphy();
+        }
     };
 
     // registers for the change event from the profile selector
@@ -799,6 +827,112 @@ jQuery(document).ready(function() {
             viewportContainer.removeClass("caret-active");
         }
         updateUrl();
+    });
+
+    // initializes the calligraphy canvas inside the viewport
+    // preview at the zoomed pixel size so that jSignature mouse
+    // coordinates match screen space, then applies a counter-scale
+    // to fit within the viewport preview transform
+    const initCalligraphy = function() {
+        if (!currentProfile) return;
+        const padding = getMargins();
+        const safeW = (currentProfile.width - padding.left - padding.right) * VIEWPORT_SCALE;
+        const safeH = (currentProfile.height - padding.top - padding.bottom) * VIEWPORT_SCALE;
+        const zoom = parseFloat(zoomRange.val()) || 1;
+        const lineWidth = currentProfile.calligraphy
+            ? currentProfile.calligraphy.line_width || 2
+            : 2;
+        calligraphyContainer.calligraphy("init", {
+            width: Math.round(safeW * zoom),
+            height: Math.round(safeH * zoom),
+            lineWidth: lineWidth * zoom
+        });
+        calligraphyContainer.css({
+            transform: "scale(" + 1 / zoom + ")",
+            "-o-transform": "scale(" + 1 / zoom + ")",
+            "-ms-transform": "scale(" + 1 / zoom + ")",
+            "-moz-transform": "scale(" + 1 / zoom + ")",
+            "-khtml-transform": "scale(" + 1 / zoom + ")",
+            "-webkit-transform": "scale(" + 1 / zoom + ")",
+            "transform-origin": "0px 0px",
+            "-o-transform-origin": "0px 0px",
+            "-ms-transform-origin": "0px 0px",
+            "-moz-transform-origin": "0px 0px",
+            "-khtml-transform-origin": "0px 0px",
+            "-webkit-transform-origin": "0px 0px"
+        });
+    };
+
+    // registers for the change in the calligraphy mode checkbox
+    // to toggle between text editing and calligraphy drawing
+    calligraphyMode.bind("change", function() {
+        const enabled = calligraphyMode.prop("checked");
+        if (enabled) {
+            viewportPreview.addClass("calligraphy-active");
+            fontsContainer.hide();
+            keyboardContainer.hide();
+            emojisContainer.hide();
+            emojispContainer.hide();
+            fontSizeContainer.removeClass("visible");
+            viewportOptionsKeyboard.hide();
+            viewportOptionsCaret.hide();
+            calligraphyControls.addClass("visible");
+            zoomRange.prop("disabled", true);
+            initCalligraphy();
+            body.data("calligraphy", null);
+        } else {
+            viewportPreview.removeClass("calligraphy-active");
+            calligraphyControls.removeClass("visible");
+            viewportOptionsKeyboard.show();
+            viewportOptionsCaret.show();
+            zoomRange.prop("disabled", false);
+            calligraphyContainer.calligraphy("reset");
+            body.data("calligraphy", null);
+            const font = body.data("font");
+            if (font) {
+                fontsContainer.show();
+                const showKeyboard = keyboardMode.prop("checked");
+                if (showKeyboard) {
+                    if (font === "Cool Emojis") {
+                        emojisContainer.show();
+                    } else if (font === "Cool Emojis Pantograph") {
+                        emojispContainer.show();
+                    } else {
+                        keyboardContainer.show();
+                    }
+                }
+            } else {
+                fontsContainer.show();
+            }
+            if (currentProfile && currentProfile.font_size) {
+                fontSizeContainer.addClass("visible");
+            }
+        }
+        updateUrl();
+    });
+
+    // registers for the change event from the calligraphy
+    // canvas to store the SVG data in the body element
+    calligraphyContainer.bind("calligraphy", function(event, hasData) {
+        if (hasData) {
+            const data = calligraphyContainer.calligraphy("data");
+            body.data("calligraphy", data);
+        } else {
+            body.data("calligraphy", null);
+        }
+    });
+
+    // registers for the click on the calligraphy undo button
+    // to remove the last stroke from the drawing
+    calligraphyUndo.click(function() {
+        calligraphyContainer.calligraphy("undo");
+    });
+
+    // registers for the click on the calligraphy clear button
+    // to reset the entire drawing canvas
+    calligraphyClear.click(function() {
+        calligraphyContainer.calligraphy("reset");
+        body.data("calligraphy", null);
     });
 
     // registers for the mouse move event on the viewport preview
@@ -984,6 +1118,8 @@ jQuery(document).ready(function() {
         if (!showGuidelines) params.set("guidelines", "0");
         const showCaret = caretMode.prop("checked");
         if (!showCaret) params.set("caret", "0");
+        const calligraphyEnabled = calligraphyMode.prop("checked");
+        if (calligraphyEnabled) params.set("calligraphy", "1");
         const fullscreen = urlParams.get("fullscreen");
         if (fullscreen === "1") params.set("fullscreen", "1");
         if (theme !== "default") params.set("theme", theme);